refactor: Separated the hooks inside Signup
