+++ src/hooks/useLoginUser.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/router";
 import { loginUserFn } from "../utils/authApi";
 import { LoginInput } from "../types/LoginInput";
 
-export default function useRegisterUser() {
+export default function useLoginUser() {
   const router = useRouter();
   const mutation = useMutation(
     (userData: LoginInput) => loginUserFn(userData),
